refactor: テストがしやすいようにリファクタリング
